# api/openapi.yaml
openapi: 3.0.0
info:
  description: Parvaeres magic deployment API
  title: Parvaeres API
  version: 0.0.1
servers:
- description: Here we go, let's make this happen!
  url: http://api.poc.parvaeres.io/v1
paths:
  /:
    description: This is Parvaeres Server
    get:
      responses:
        "200":
          description: OK
      summary: Alive
    summary: Parvaeres Server
  /deployment:
    description: This resource represents the deployments in the system
    get:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetDeploymentRequest'
        description: Additional parameters for request
        required: false
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDeploymentResponse'
          description: OK
      summary: Get all deployments
    post:
      parameters:
      - description: a very rudimentary security token
        explode: false
        in: header
        name: parvaerestoken
        required: false
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeploymentRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreateDeploymentRequest'
        description: Deployment parameters
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDeploymentResponse'
          description: Accepted
      summary: Create a new deployment
    summary: Deployments
  /deployment/{deploymentId}:
    delete:
      parameters:
      - description: the id of the deployment to delete
        explode: false
        in: path
        name: deploymentId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentStatus'
          description: Accepted
      summary: Delete deployment with id deploymentId
    description: This resource represents an individual deployment
    get:
      parameters:
      - description: the id of the deployment to retrieve
        explode: false
        in: path
        name: deploymentId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentStatus'
          description: OK
      summary: Get the deployment with id deploymentId
    summary: A deployment
  /deployment/{deploymentId}/logs:
    description: Retrieve the logs of the deployment
    get:
      parameters:
      - description: the id of the deployment for which retrieve the logs
        explode: false
        in: path
        name: deploymentId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDeploymentLogsResponse'
          description: OK
      summary: Get the logs of the deployment with deploymentId
    summary: Get logs
components:
  schemas:
    CreateDeploymentRequest:
      example:
        Path: Path
        Repository: Repository
        Email: Email
      properties:
        Email:
          type: string
        Repository:
          type: string
        Path:
          type: string
      type: object
    CreateDeploymentResponse:
      example:
        Message: Message
        Error: true
        Items:
        - '{}'
        - '{}'
      properties:
        Message:
          type: string
        Error:
          type: boolean
        Items:
          items:
            allOf:
            - $ref: '#/components/schemas/DeploymentStatus'
            type: object
          type: array
      type: object
    GetDeploymentRequest:
      example:
        Filter: Filter
        UUID: UUID
      properties:
        UUID:
          type: string
        Filter:
          type: string
      type: object
    GetDeploymentResponse:
      example:
        Message: Message
        Error: true
        Items:
        - '{}'
        - '{}'
      properties:
        Message:
          type: string
        Error:
          type: boolean
        Items:
          items:
            allOf:
            - $ref: '#/components/schemas/DeploymentStatus'
            type: object
          type: array
      type: object
    GetDeploymentLogsResponse:
      example:
        Message: Message
        Error: true
        Items:
        - '{}'
        - '{}'
      properties:
        Message:
          type: string
        Error:
          type: boolean
        Items:
          items:
            allOf:
            - $ref: '#/components/schemas/Logs'
            type: object
          type: array
      type: object
    DeploymentStatus:
      example:
        Path: Path
        Errors:
        - Errors
        - Errors
        Email: Email
        LogsURL: LogsURL
        LiveURLs:
        - LiveURLs
        - LiveURLs
        RepoURL: RepoURL
        UUID: UUID
      properties:
        UUID:
          type: string
        RepoURL:
          type: string
        Path:
          type: string
        Email:
          type: string
        Status:
          $ref: '#/components/schemas/DeploymentStatusType'
        LiveURLs:
          items:
            type: string
          type: array
        Errors:
          items:
            type: string
          type: array
        LogsURL:
          type: string
      type: object
    DeploymentStatusType:
      enum:
      - PENDING
      - SYNCING
      - DEPLOYED
      - ERROR
      - UNKNOWN
      type: string
    Logs:
      properties:
        Pod:
          type: string
        Container:
          type: string
        Logs:
          type: string
      type: object